I did some things with the narrative system. You should play sounds onto the narrationSphere things. I also made WorldGlobal load all the sounds in the folder with names proper and junk :D Could do it for the other things, but whatevs!
There is also and example of how to use yield in C# for Unity.
I feel like it will be damn useful for sending events when the audio is finished playing.
Still have no idea how the hell we are going to be able to do more things efficiently based on when audio cues happen.
Maybe we can yield until the needed method returns a value or something.